requirements.txt
requests==2.28.2
pandas==1.5.3
numpy==1.24.2
google==3.0.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
openpyxl==3.1.2
python-dotenv==1.0.0
